refactor: move domain errors to domain folder
