+++ src/domain/entities/CPF.ts
@@ -1,4 +1,4 @@
-import { InvalidCPFerror } from './errors'
+import { InvalidCPFerror } from '../errors'
 
 export class CPF {
   public value: string
