+++ src/domain/errors/empty-properties/InvalidEmptyProperty.ts
