+++ src/domain/errors/empty-properties/index.ts
