+++ src/domain/errors/entity-not-found/CouponNotFound.ts
