+++ src/domain/errors/entity-not-found/EntityNotFound.ts
