+++ src/domain/errors/entity-not-found/ItemNotFound.ts
