+++ src/domain/errors/entity-not-found/index.ts
