+++ src/domain/errors/index.ts
@@ -1,3 +1,4 @@
 export * from './empty-properties'
+export * from './entity-not-found'
 export * from './ExpiredCoupon'
 export * from './InvalidCPF'
