+++ tests/application/use-cases/PlaceOrder.spec.ts
@@ -1,6 +1,7 @@
+import { ItemRepository, OrderRepository } from '@/domain/repositories'
 import { PlaceOrder } from '@/application/use-cases'
 import { PlaceOrderInput } from '@/application/dtos/place-order'
-import { ItemRepository, OrderRepository } from '@/domain/repositories'
+
 import { ItemRepositoryStub, OrderRepositoryStub } from '@/tests/stub/repositories'
 
 type makeSutTypes = {
