+++ tests/infra/ItemRepository.test.ts
@@ -1,10 +1,10 @@
 import { randomUUID } from 'crypto'
 
 import { Item } from '@/domain/entities'
+import { ItemNotFoundError } from '@/domain/errors'
 
 import { ItemRepositoryPostgres } from '@/infra/repositories'
 import { DatabaseConnectionAdapter, DatabaseConnection } from '@/infra/database'
-import { ItemNotFoundError } from '@/infra/errors'
 
 type makeSutType = {
   itemRepository: ItemRepositoryPostgres
